requirements.txt
absl-py==1.3.0
appdirs==1.4.4
astunparse==1.6.3
audioread==3.0.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cycler==0.11.0
decorator==5.1.1
flatbuffers==22.9.24
fonttools==4.36.0
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.3
importlib-metadata==5.0.0
joblib==1.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
librosa==0.9.2
llvmlite==0.39.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numba==0.56.3
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pooch==1.6.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
six==1.16.0
SoundFile==0.10.3.post1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
termcolor==2.0.1
threadpoolctl==3.1.0
tk==0.1.0
typing_extensions==4.4.0
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
